Add files via upload
synth.semf.py -> a python script that synthesizes the binding energy per nucleon from the semi-empirical mass formula.

semf.settings.inp -> an input file for the above script. you can control the settings of the code with this file or change the flags in the script itself!

sr.semf.py -> a symbolic regression script that will attempt to recreate the semi-empirical mass formula given the data from the synth.semf.py.